+++ test/unit/resourceSpec.js
@@ -4,7 +4,7 @@ var chai = require('chai');
 var expect = chai.expect;
 var sinon = require('sinon');
 var sinonChai = require('sinon-chai');
-var Resource = require('../../services/resourceService');
+var Resource = require('../../services/resource');
 
 chai.use(sinonChai);
